ANITA LOOK DESCRIPTION
so idk what causing the background around the image to be a different colour... its the samee colour on the photoshop file and when its saved but for some reason when its published in the code its a different colour slightly and u can see it for some reason. idk why unless its something to do with the card that you've put down but i dont want to interfere with the code!